requirements.txt
easyocr
opencv-python
Pillow
torch
transformers
sentencepiece
timm
rapidfuzz
